all 7 stages implemented for the new release
