+++ backend/src/api/admin.rs
@@ -503,6 +503,7 @@ async fn delete_retention_policy(
 // ============ Plugins - RustChat Calls Plugin ============
 
 #[derive(Debug, serde::Serialize, serde::Deserialize)]
+#[allow(dead_code)]
 pub struct CallsPluginConfig {
     pub enabled: bool,
     pub turn_server_enabled: bool,
