+++ backend/src/api/v4/calls_plugin/sfu/tracks.rs
@@ -1,6 +1,7 @@
 //! Track Management for SFU
 //!
 //! Manages audio/video tracks from participants and handles track forwarding.
+#![allow(dead_code)]
 
 use std::collections::HashMap;
 use std::sync::Arc;
