+++ backend/src/api/v4/posts.rs
@@ -1,12 +1,10 @@
 use axum::{
     body::Bytes,
-    extract::{Path, Query, State},
-    http::StatusCode,
+    extract::{Path, State},
     response::IntoResponse,
     routing::{delete, get, post, put},
     Json, Router,
 };
-use chrono::{DateTime, Utc};
 use serde::Deserialize;
 use serde_json::json;
 use uuid::Uuid;
@@ -21,6 +19,17 @@ use crate::mattermost_compat::{
 use crate::models::{CreatePost, FileInfo};
 use crate::realtime::{EventType, WsBroadcast, WsEnvelope};
 use crate::services::posts;
+
+mod reactions;
+mod search;
+mod unread;
+
+pub(crate) use reactions::reactions_for_posts;
+use reactions::{add_reaction, get_reactions, remove_reaction, remove_reaction_for_user};
+use search::{search_posts_all_teams, search_team_posts};
+use unread::{
+    delete_acknowledgement_for_post, get_posts_around_unread, save_acknowledgement_for_post,
+};
 
 pub fn router() -> Router<AppState> {
     Router::new()
@@ -533,7 +542,8 @@ async fn handle_post_action(
 
 #[derive(Deserialize)]
 struct MovePostRequest {
-    channel_id: String,
+    #[serde(rename = "channel_id")]
+    _channel_id: String,
 }
 
 async fn move_post(
@@ -903,276 +913,6 @@ async fn patch_post(
     Ok(Json(updated.into()))
 }
 
-#[derive(Deserialize)]
-struct ReactionRequest {
-    user_id: String,
-    post_id: String,
-    emoji_name: String,
-}
-
-async fn add_reaction(
-    State(state): State<AppState>,
-    auth: MmAuthUser,
-    headers: axum::http::HeaderMap,
-    body: Bytes,
-) -> ApiResult<(StatusCode, Json<mm::Reaction>)> {
-    let input: ReactionRequest = parse_body(&headers, &body, "Invalid reaction body")?;
-    let input_user_id = parse_mm_or_uuid(&input.user_id)
-        .ok_or_else(|| AppError::Validation("Invalid user_id".to_string()))?;
-    if input_user_id != auth.user_id {
-        return Err(AppError::Forbidden(
-            "Cannot react for other user".to_string(),
-        ));
-    }
-
-    let post_id = parse_mm_or_uuid(&input.post_id)
-        .ok_or_else(|| AppError::Validation("Invalid post_id".to_string()))?;
-
-    // Normalize emoji name (e.g., convert Unicode character to short name)
-    let emoji_name =
-        crate::mattermost_compat::emoji_data::get_short_name_for_emoji(&input.emoji_name);
-
-    // Validate emoji name
-    if !crate::mattermost_compat::emoji_data::is_valid_emoji_name(&emoji_name) {
-        return Err(AppError::BadRequest("Invalid emoji name".to_string()));
-    }
-
-    // Check if it exists as either a system emoji or custom emoji
-    if !crate::mattermost_compat::emoji_data::is_system_emoji(&emoji_name) {
-        let exists: bool = sqlx::query_scalar(
-            "SELECT EXISTS(SELECT 1 FROM custom_emojis WHERE name = $1 AND delete_at IS NULL)",
-        )
-        .bind(&emoji_name)
-        .fetch_one(&state.db)
-        .await?;
-
-        if !exists {
-            return Err(AppError::NotFound("Emoji not found".to_string()));
-        }
-    }
-
-    let reaction: crate::models::post::Reaction = sqlx::query_as(
-        r#"
-        INSERT INTO reactions (user_id, post_id, emoji_name)
-        VALUES ($1, $2, $3)
-        ON CONFLICT (user_id, post_id, emoji_name) DO UPDATE SET emoji_name = $3
-        RETURNING *
-        "#,
-    )
-    .bind(auth.user_id)
-    .bind(post_id)
-    .bind(&emoji_name)
-    .fetch_one(&state.db)
-    .await?;
-
-    let channel_id: Uuid = sqlx::query_scalar("SELECT channel_id FROM posts WHERE id = $1")
-        .bind(post_id)
-        .fetch_one(&state.db)
-        .await?;
-
-    let mm_reaction = mm::Reaction {
-        user_id: encode_mm_id(reaction.user_id),
-        post_id: encode_mm_id(reaction.post_id),
-        emoji_name: reaction.emoji_name,
-        create_at: reaction.created_at.timestamp_millis(),
-        update_at: reaction.created_at.timestamp_millis(),
-        delete_at: 0,
-        channel_id: encode_mm_id(channel_id),
-        remote_id: "".to_string(),
-    };
-
-    let reaction_json = serde_json::to_string(&mm_reaction).unwrap();
-    let data = serde_json::json!({ "reaction": reaction_json });
-
-    let broadcast = WsEnvelope::event(EventType::ReactionAdded, data, Some(channel_id))
-        .with_broadcast(WsBroadcast {
-            channel_id: Some(channel_id),
-            team_id: None,
-            user_id: None,
-            exclude_user_id: None,
-        });
-    state.ws_hub.broadcast(broadcast).await;
-
-    Ok((StatusCode::CREATED, Json(mm_reaction)))
-}
-
-pub(crate) async fn reactions_for_posts(
-    state: &AppState,
-    post_ids: &[Uuid],
-) -> ApiResult<std::collections::HashMap<Uuid, Vec<mm::Reaction>>> {
-    use std::collections::HashMap;
-
-    if post_ids.is_empty() {
-        return Ok(HashMap::new());
-    }
-
-    let reactions: Vec<(Uuid, Uuid, String, DateTime<Utc>, Uuid)> = sqlx::query_as(
-        r#"
-        SELECT r.post_id, r.user_id, r.emoji_name, r.created_at, p.channel_id
-        FROM reactions r
-        JOIN posts p ON p.id = r.post_id
-        WHERE r.post_id = ANY($1)
-        "#,
-    )
-    .bind(post_ids)
-    .fetch_all(&state.db)
-    .await?;
-
-    let mut map: HashMap<Uuid, Vec<mm::Reaction>> = HashMap::new();
-    for (post_id, user_id, emoji_name, created_at, channel_id) in reactions {
-        map.entry(post_id).or_default().push(mm::Reaction {
-            user_id: encode_mm_id(user_id),
-            post_id: encode_mm_id(post_id),
-            emoji_name: crate::mattermost_compat::emoji_data::get_short_name_for_emoji(&emoji_name),
-            create_at: created_at.timestamp_millis(),
-            update_at: created_at.timestamp_millis(),
-            delete_at: 0,
-            channel_id: encode_mm_id(channel_id),
-            remote_id: "".to_string(),
-        });
-    }
-
-    Ok(map)
-}
-
-async fn remove_reaction(
-    State(state): State<AppState>,
-    auth: MmAuthUser,
-    Path((post_id, emoji_name)): Path<(String, String)>,
-) -> ApiResult<impl IntoResponse> {
-    let post_id = parse_mm_or_uuid(&post_id)
-        .ok_or_else(|| AppError::BadRequest("Invalid post_id".to_string()))?;
-
-    remove_reaction_internal(&state, auth.user_id, post_id, &emoji_name).await?;
-
-    Ok(Json(serde_json::json!({"status": "OK"})))
-}
-
-async fn remove_reaction_for_user(
-    State(state): State<AppState>,
-    auth: MmAuthUser,
-    Path((user_id, post_id, emoji_name)): Path<(String, String, String)>,
-) -> ApiResult<impl IntoResponse> {
-    let resolved_user_id = if user_id == "me" {
-        auth.user_id
-    } else {
-        parse_mm_or_uuid(&user_id)
-            .ok_or_else(|| AppError::BadRequest("Invalid user_id".to_string()))?
-    };
-
-    if resolved_user_id != auth.user_id {
-        return Err(AppError::Forbidden(
-            "Cannot remove reactions for other user".to_string(),
-        ));
-    }
-
-    let post_id = parse_mm_or_uuid(&post_id)
-        .ok_or_else(|| AppError::BadRequest("Invalid post_id".to_string()))?;
-
-    remove_reaction_internal(&state, resolved_user_id, post_id, &emoji_name).await?;
-
-    Ok(Json(serde_json::json!({"status": "OK"})))
-}
-
-async fn remove_reaction_internal(
-    state: &AppState,
-    user_id: Uuid,
-    post_id: Uuid,
-    emoji_name: &str,
-) -> ApiResult<()> {
-    // Normalize emoji name (e.g., convert Unicode character to short name)
-    let emoji_name = crate::mattermost_compat::emoji_data::get_short_name_for_emoji(emoji_name);
-
-    let reaction: Option<crate::models::post::Reaction> = sqlx::query_as(
-        "SELECT * FROM reactions WHERE user_id = $1 AND post_id = $2 AND emoji_name = $3",
-    )
-    .bind(user_id)
-    .bind(post_id)
-    .bind(&emoji_name)
-    .fetch_optional(&state.db)
-    .await?;
-
-    if let Some(r) = reaction {
-        sqlx::query(
-            "DELETE FROM reactions WHERE user_id = $1 AND post_id = $2 AND emoji_name = $3",
-        )
-        .bind(user_id)
-        .bind(post_id)
-        .bind(&emoji_name)
-        .execute(&state.db)
-        .await?;
-
-        let channel_id: Uuid = sqlx::query_scalar("SELECT channel_id FROM posts WHERE id = $1")
-            .bind(post_id)
-            .fetch_one(&state.db)
-            .await?;
-
-        let mm_reaction = mm::Reaction {
-            user_id: encode_mm_id(r.user_id),
-            post_id: encode_mm_id(r.post_id),
-            emoji_name: r.emoji_name,
-            create_at: r.created_at.timestamp_millis(),
-            update_at: r.created_at.timestamp_millis(),
-            delete_at: 0,
-            channel_id: encode_mm_id(channel_id),
-            remote_id: "".to_string(),
-        };
-
-        let reaction_json = serde_json::to_string(&mm_reaction).unwrap();
-        let data = serde_json::json!({ "reaction": reaction_json });
-
-        let broadcast = WsEnvelope::event(EventType::ReactionRemoved, data, Some(channel_id))
-            .with_broadcast(WsBroadcast {
-                channel_id: Some(channel_id),
-                team_id: None,
-                user_id: None,
-                exclude_user_id: None,
-            });
-        state.ws_hub.broadcast(broadcast).await;
-    }
-
-    Ok(())
-}
-
-async fn get_reactions(
-    State(state): State<AppState>,
-    Path(post_id): Path<String>,
-) -> ApiResult<Json<Vec<mm::Reaction>>> {
-    let post_id = parse_mm_or_uuid(&post_id)
-        .ok_or_else(|| AppError::BadRequest("Invalid post_id".to_string()))?;
-    let reactions: Vec<(Uuid, Uuid, String, DateTime<Utc>, Uuid)> = sqlx::query_as(
-        r#"
-        SELECT r.user_id, r.post_id, r.emoji_name, r.created_at, p.channel_id
-        FROM reactions r
-        JOIN posts p ON p.id = r.post_id
-        WHERE r.post_id = $1
-        "#,
-    )
-    .bind(post_id)
-    .fetch_all(&state.db)
-    .await?;
-
-    let mm_reactions = reactions
-        .into_iter()
-        .map(
-            |(user_id, post_id, emoji_name, created_at, channel_id)| mm::Reaction {
-                user_id: encode_mm_id(user_id),
-                post_id: encode_mm_id(post_id),
-                emoji_name: crate::mattermost_compat::emoji_data::get_short_name_for_emoji(
-                    &emoji_name,
-                ),
-                create_at: created_at.timestamp_millis(),
-                update_at: created_at.timestamp_millis(),
-                delete_at: 0,
-                channel_id: encode_mm_id(channel_id),
-                remote_id: "".to_string(),
-            },
-        )
-        .collect();
-
-    Ok(Json(mm_reactions))
-}
-
 /// POST /posts/{post_id}/ack - Acknowledge a post (push notification receipt)
 #[derive(Deserialize)]
 #[allow(dead_code)]
@@ -1505,432 +1245,4 @@ async fn set_post_reminder(
     .await?;
 
     Ok(Json(serde_json::json!({"status": "OK"})))
-}
-
-// ============================================================================
-// Team Posts Search
-// ============================================================================
-
-#[derive(Deserialize)]
-struct SearchPostsRequest {
-    terms: String,
-    is_or_search: bool,
-    #[serde(default)]
-    time_zone_offset: i32,
-    #[serde(default)]
-    include_deleted_channels: bool,
-    #[serde(default)]
-    page: i32,
-    #[serde(default = "default_per_page")]
-    per_page: i32,
-}
-
-fn default_per_page() -> i32 {
-    60
-}
-
-/// POST /api/v4/teams/{team_id}/posts/search - Search posts in team
-async fn search_team_posts(
-    State(state): State<AppState>,
-    auth: MmAuthUser,
-    Path(team_id): Path<String>,
-    headers: axum::http::HeaderMap,
-    body: Bytes,
-) -> ApiResult<Json<mm::PostListWithSearchMatches>> {
-    let team_id = parse_mm_or_uuid(&team_id)
-        .ok_or_else(|| AppError::BadRequest("Invalid team_id".to_string()))?;
-
-    let input: SearchPostsRequest = parse_body(&headers, &body, "Invalid search body")?;
-
-    // Verify user is member of team
-    let _: crate::models::TeamMember =
-        sqlx::query_as("SELECT * FROM team_members WHERE team_id = $1 AND user_id = $2")
-            .bind(team_id)
-            .bind(auth.user_id)
-            .fetch_optional(&state.db)
-            .await?
-            .ok_or_else(|| AppError::Forbidden("Not a member of this team".to_string()))?;
-
-    let limit = input.per_page.min(200).max(1) as i64;
-    let offset = (input.page.max(0) as i64) * limit;
-
-    // Build search query using ILIKE for simple text search
-    let search_terms = format!("%{}%", input.terms.replace('%', "\\%").replace('_', "\\_"));
-
-    let posts: Vec<crate::models::post::PostResponse> = sqlx::query_as(
-        r#"
-        SELECT p.id, p.channel_id, p.user_id, p.root_post_id, p.message, p.props, p.file_ids,
-               p.is_pinned, p.created_at, p.edited_at, p.deleted_at,
-               p.reply_count::int8 as reply_count,
-               p.last_reply_at, p.seq,
-               u.username, u.avatar_url, u.email
-        FROM posts p
-        LEFT JOIN users u ON p.user_id = u.id
-        JOIN channels c ON p.channel_id = c.id
-        JOIN channel_members cm ON cm.channel_id = c.id AND cm.user_id = $4
-        WHERE c.team_id = $1
-          AND p.message ILIKE $2
-          AND p.deleted_at IS NULL
-        ORDER BY p.created_at DESC
-        LIMIT $3 OFFSET $5
-        "#,
-    )
-    .bind(team_id)
-    .bind(&search_terms)
-    .bind(limit)
-    .bind(auth.user_id)
-    .bind(offset)
-    .fetch_all(&state.db)
-    .await?;
-
-    let mut order = Vec::new();
-    let mut posts_map: std::collections::HashMap<String, mm::Post> =
-        std::collections::HashMap::new();
-    let mut matches_map: std::collections::HashMap<String, Vec<String>> =
-        std::collections::HashMap::new();
-    let mut post_ids = Vec::new();
-    let mut id_map = Vec::new();
-
-    for p in posts {
-        let id = encode_mm_id(p.id);
-        post_ids.push(p.id);
-        id_map.push((p.id, id.clone()));
-        order.push(id.clone());
-        // Simple match highlighting - find term positions
-        let match_positions: Vec<String> = vec![];
-        matches_map.insert(id.clone(), match_positions);
-        posts_map.insert(id, p.into());
-    }
-
-    let reactions_map = reactions_for_posts(&state, &post_ids).await?;
-    for (post_uuid, post_id) in id_map {
-        if let Some(reactions) = reactions_map.get(&post_uuid) {
-            if !reactions.is_empty() {
-                if let Some(post) = posts_map.get_mut(&post_id) {
-                    post.metadata = Some(json!({ "reactions": reactions }));
-                }
-            }
-        }
-    }
-
-    Ok(Json(mm::PostListWithSearchMatches {
-        order,
-        posts: posts_map,
-        matches: Some(matches_map),
-        next_post_id: String::new(),
-        prev_post_id: String::new(),
-    }))
-}
-
-/// POST /api/v4/posts/search - Search posts across all teams
-async fn search_posts_all_teams(
-    State(state): State<AppState>,
-    auth: MmAuthUser,
-    headers: axum::http::HeaderMap,
-    body: Bytes,
-) -> ApiResult<Json<mm::PostListWithSearchMatches>> {
-    let input: SearchPostsRequest = parse_body(&headers, &body, "Invalid search body")?;
-
-    let limit = input.per_page.min(200).max(1) as i64;
-    let offset = (input.page.max(0) as i64) * limit;
-    let search_terms = format!("%{}%", input.terms.replace('%', "\\%").replace('_', "\\_"));
-
-    let posts: Vec<crate::models::post::PostResponse> = sqlx::query_as(
-        r#"
-        SELECT p.id, p.channel_id, p.user_id, p.root_post_id, p.message, p.props, p.file_ids,
-               p.is_pinned, p.created_at, p.edited_at, p.deleted_at,
-               p.reply_count::int8 as reply_count,
-               p.last_reply_at, p.seq,
-               u.username, u.avatar_url, u.email
-        FROM posts p
-        LEFT JOIN users u ON p.user_id = u.id
-        JOIN channel_members cm ON cm.channel_id = p.channel_id AND cm.user_id = $2
-        WHERE p.message ILIKE $1
-          AND p.deleted_at IS NULL
-        ORDER BY p.created_at DESC
-        LIMIT $3 OFFSET $4
-        "#,
-    )
-    .bind(&search_terms)
-    .bind(auth.user_id)
-    .bind(limit)
-    .bind(offset)
-    .fetch_all(&state.db)
-    .await?;
-
-    let mut order = Vec::new();
-    let mut posts_map: std::collections::HashMap<String, mm::Post> =
-        std::collections::HashMap::new();
-    let mut matches_map: std::collections::HashMap<String, Vec<String>> =
-        std::collections::HashMap::new();
-    let mut post_ids = Vec::new();
-    let mut id_map = Vec::new();
-
-    for p in posts {
-        let id = encode_mm_id(p.id);
-        post_ids.push(p.id);
-        id_map.push((p.id, id.clone()));
-        order.push(id.clone());
-        matches_map.insert(id.clone(), Vec::new());
-        posts_map.insert(id, p.into());
-    }
-
-    let reactions_map = reactions_for_posts(&state, &post_ids).await?;
-    for (post_uuid, post_id) in id_map {
-        if let Some(reactions) = reactions_map.get(&post_uuid) {
-            if !reactions.is_empty() {
-                if let Some(post) = posts_map.get_mut(&post_id) {
-                    post.metadata = Some(json!({ "reactions": reactions }));
-                }
-            }
-        }
-    }
-
-    Ok(Json(mm::PostListWithSearchMatches {
-        order,
-        posts: posts_map,
-        matches: Some(matches_map),
-        next_post_id: String::new(),
-        prev_post_id: String::new(),
-    }))
-}
-
-// ============================================================================
-// Posts Around Unread
-// ============================================================================
-
-#[derive(Deserialize)]
-struct PostsUnreadQuery {
-    #[serde(default = "default_limit")]
-    limit_before: i32,
-    #[serde(default = "default_limit")]
-    limit_after: i32,
-    #[serde(rename = "skipFetchThreads", default)]
-    skip_fetch_threads: bool,
-    #[serde(rename = "collapsedThreads", default)]
-    collapsed_threads: bool,
-    #[serde(rename = "collapsedThreadsExtended", default)]
-    collapsed_threads_extended: bool,
-}
-
-fn default_limit() -> i32 {
-    60
-}
-
-#[derive(Deserialize)]
-struct PostsUnreadPath {
-    user_id: String,
-    channel_id: String,
-}
-
-/// GET /api/v4/users/{user_id}/channels/{channel_id}/posts/unread
-async fn get_posts_around_unread(
-    State(state): State<AppState>,
-    auth: MmAuthUser,
-    Path(path): Path<PostsUnreadPath>,
-    Query(query): Query<PostsUnreadQuery>,
-) -> ApiResult<Json<mm::PostList>> {
-    let user_id = super::users::resolve_user_id(&path.user_id, &auth)
-        .map_err(|_| AppError::Forbidden("Cannot access another user's posts".to_string()))?;
-
-    let channel_id = parse_mm_or_uuid(&path.channel_id)
-        .ok_or_else(|| AppError::BadRequest("Invalid channel_id".to_string()))?;
-
-    // Verify membership
-    let _: crate::models::ChannelMember =
-        sqlx::query_as("SELECT * FROM channel_members WHERE channel_id = $1 AND user_id = $2")
-            .bind(channel_id)
-            .bind(user_id)
-            .fetch_optional(&state.db)
-            .await?
-            .ok_or_else(|| AppError::Forbidden("Not a member of this channel".to_string()))?;
-
-    let limit_before = query.limit_before.min(200).max(0) as i64;
-    let limit_after = query.limit_after.min(200).max(1) as i64;
-
-    // Find the oldest unread post using channel_reads
-    let last_read_seq: Option<i64> = sqlx::query_scalar(
-        "SELECT last_read_message_id FROM channel_reads WHERE user_id = $1 AND channel_id = $2",
-    )
-    .bind(user_id)
-    .bind(channel_id)
-    .fetch_optional(&state.db)
-    .await?
-    .flatten();
-
-    let last_read_seq = last_read_seq.unwrap_or(0);
-
-    // Get posts around the unread point
-    let posts: Vec<crate::models::post::PostResponse> = sqlx::query_as(
-        r#"
-        (
-            SELECT p.id, p.channel_id, p.user_id, p.root_post_id, p.message, p.props, p.file_ids,
-                   p.is_pinned, p.created_at, p.edited_at, p.deleted_at,
-                   p.reply_count::int8 as reply_count,
-                   p.last_reply_at, p.seq,
-                   u.username, u.avatar_url, u.email
-            FROM posts p
-            LEFT JOIN users u ON p.user_id = u.id
-            WHERE p.channel_id = $1 AND p.seq <= $2 AND p.deleted_at IS NULL
-            ORDER BY p.seq DESC
-            LIMIT $3
-        )
-        UNION ALL
-        (
-            SELECT p.id, p.channel_id, p.user_id, p.root_post_id, p.message, p.props, p.file_ids,
-                   p.is_pinned, p.created_at, p.edited_at, p.deleted_at,
-                   p.reply_count::int8 as reply_count,
-                   p.last_reply_at, p.seq,
-                   u.username, u.avatar_url, u.email
-            FROM posts p
-            LEFT JOIN users u ON p.user_id = u.id
-            WHERE p.channel_id = $1 AND p.seq > $2 AND p.deleted_at IS NULL
-            ORDER BY p.seq ASC
-            LIMIT $4
-        )
-        ORDER BY seq DESC
-        "#,
-    )
-    .bind(channel_id)
-    .bind(last_read_seq)
-    .bind(limit_before)
-    .bind(limit_after)
-    .fetch_all(&state.db)
-    .await?;
-
-    let mut order = Vec::new();
-    let mut posts_map: std::collections::HashMap<String, mm::Post> =
-        std::collections::HashMap::new();
-    let mut post_ids = Vec::new();
-    let mut id_map = Vec::new();
-
-    for p in posts {
-        let id = encode_mm_id(p.id);
-        post_ids.push(p.id);
-        id_map.push((p.id, id.clone()));
-        order.push(id.clone());
-        posts_map.insert(id, p.into());
-    }
-
-    let reactions_map = reactions_for_posts(&state, &post_ids).await?;
-    for (post_uuid, post_id) in id_map {
-        if let Some(reactions) = reactions_map.get(&post_uuid) {
-            if !reactions.is_empty() {
-                if let Some(post) = posts_map.get_mut(&post_id) {
-                    post.metadata = Some(json!({ "reactions": reactions }));
-                }
-            }
-        }
-    }
-
-    Ok(Json(mm::PostList {
-        order,
-        posts: posts_map,
-        next_post_id: String::new(),
-        prev_post_id: String::new(),
-    }))
-}
-
-// ============================================================================
-// Post Acknowledgements
-// ============================================================================
-
-#[derive(Deserialize)]
-struct AckPath {
-    user_id: String,
-    post_id: String,
-}
-
-/// POST /api/v4/users/{user_id}/posts/{post_id}/ack - Acknowledge a post
-async fn save_acknowledgement_for_post(
-    State(state): State<AppState>,
-    auth: MmAuthUser,
-    Path(path): Path<AckPath>,
-) -> ApiResult<Json<mm::PostAcknowledgement>> {
-    let user_id = super::users::resolve_user_id(&path.user_id, &auth)
-        .map_err(|_| AppError::Forbidden("Cannot acknowledge for another user".to_string()))?;
-
-    let post_id = parse_mm_or_uuid(&path.post_id)
-        .ok_or_else(|| AppError::BadRequest("Invalid post_id".to_string()))?;
-
-    // Verify user can access the post
-    let channel_id: Uuid =
-        sqlx::query_scalar("SELECT channel_id FROM posts WHERE id = $1 AND deleted_at IS NULL")
-            .bind(post_id)
-            .fetch_optional(&state.db)
-            .await?
-            .ok_or_else(|| AppError::NotFound("Post not found".to_string()))?;
-
-    let _: crate::models::ChannelMember =
-        sqlx::query_as("SELECT * FROM channel_members WHERE channel_id = $1 AND user_id = $2")
-            .bind(channel_id)
-            .bind(user_id)
-            .fetch_optional(&state.db)
-            .await?
-            .ok_or_else(|| AppError::Forbidden("Not a member of this channel".to_string()))?;
-
-    let now = chrono::Utc::now();
-
-    sqlx::query(
-        r#"
-        INSERT INTO post_acknowledgements (user_id, post_id, acknowledged_at)
-        VALUES ($1, $2, $3)
-        ON CONFLICT (user_id, post_id) DO UPDATE SET acknowledged_at = $3
-        "#,
-    )
-    .bind(user_id)
-    .bind(post_id)
-    .bind(now)
-    .execute(&state.db)
-    .await?;
-
-    Ok(Json(mm::PostAcknowledgement {
-        user_id: encode_mm_id(user_id),
-        post_id: encode_mm_id(post_id),
-        acknowledged_at: now.timestamp_millis(),
-    }))
-}
-
-/// DELETE /api/v4/users/{user_id}/posts/{post_id}/ack - Delete a post acknowledgement
-async fn delete_acknowledgement_for_post(
-    State(state): State<AppState>,
-    auth: MmAuthUser,
-    Path(path): Path<AckPath>,
-) -> ApiResult<impl IntoResponse> {
-    let user_id = super::users::resolve_user_id(&path.user_id, &auth).map_err(|_| {
-        AppError::Forbidden("Cannot delete acknowledgement for another user".to_string())
-    })?;
-
-    let post_id = parse_mm_or_uuid(&path.post_id)
-        .ok_or_else(|| AppError::BadRequest("Invalid post_id".to_string()))?;
-
-    // Check if acknowledgement exists and was made within 5 minutes
-    let ack_time: Option<chrono::DateTime<chrono::Utc>> = sqlx::query_scalar(
-        "SELECT acknowledged_at FROM post_acknowledgements WHERE user_id = $1 AND post_id = $2",
-    )
-    .bind(user_id)
-    .bind(post_id)
-    .fetch_optional(&state.db)
-    .await?
-    .flatten();
-
-    if let Some(ack_time) = ack_time {
-        let now = chrono::Utc::now();
-        let five_minutes = chrono::Duration::minutes(5);
-        if now - ack_time > five_minutes {
-            return Err(AppError::Forbidden(
-                "Cannot delete acknowledgement after 5 minutes".to_string(),
-            ));
-        }
-    } else {
-        return Err(AppError::NotFound("Acknowledgement not found".to_string()));
-    }
-
-    sqlx::query("DELETE FROM post_acknowledgements WHERE user_id = $1 AND post_id = $2")
-        .bind(user_id)
-        .bind(post_id)
-        .execute(&state.db)
-        .await?;
-
-    Ok(status_ok())
 }
